refactor: collapse ConfigBuilder into Config
Move the flag-to-Config lifecycle out of ConfigBuilder.{h,cpp} into
Config.cpp's anonymous namespace, called directly from
Config(flags, isCLI). ConfigBuilder had zero non-test production
callers; the entry point everywhere is the Config constructor.

Promote cross-field invariants into Config::adaptDefaults():
regularized->recordedTeleportation, noInfomap->numTrials=1,
verbosity->!prettyOutput, printAllTrials guard, library output
defaults, output-directory normalization, and output-name default.
Library users who mutate fields and call adaptDefaults() now get
the same coherence guarantees CLI users got, closing a latent
divergence between the two construction paths.

Lifecycle keeps only the I/O steps: filesystem write check and
Log initialization. Move the post-parse interpretation of the
remaining staging slots (flow model, deprecated alias, optional
output dir) from ParameterCatalog.cpp into Config.cpp's lifecycle.
ParameterCatalog's public surface sheds applyParsedParameters.

Write Config::parsedOptions directly from
ProgramInterface::getUsedOptionArguments() and drop the
ParsedParameterSet::usedOptions intermediate. ParameterCatalog
no longer needs ParsedOption.h.

Config.h signature surface is unchanged. SWIG-published methods
and fields keep identical declarations; adaptDefaults() body
grows. ParameterCatalog's registrar layer is deliberately untouched.
